requirements.txt
channels==4.0.0
channels-rabbitmq==4.0.1
channels-redis==4.1.0
cloudinary==1.37.0
daphne==4.0.0
dj-database-url==2.1.0
Django==5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-heroku==0.3.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
oauthlib==3.2.2
Pillow==10.1.0
python-decouple==3.8
redis==5.0.1
